[FIX] resolve conflicts
